+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.2.0 - 2017-07-12
+
+### Added
+
+* Errors and Exceptions are now posted to Rollbar when they occur
+
+### Improved
+
+* The `accessToken` and `clientAccessToken` settings can now be set from the CP as well as the plugin's config file.
+
 ## 0.1.0 - 2017-07-12
 
 * Initial plugin with a `Settings` model.